pyproject.toml
[tool.poetry]
name = "financial-assistant"
version = "0.1.0"
description = ""
authors = ["Ivan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.10"
aiosignal = "1.3.1"
altair = "5.4.1"
async-timeout = "4.0.3"
attrs = "24.2.0"
beautifulsoup4 = "4.12.3"
blinker = "1.8.2"
cachetools = "5.5.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
click = "8.1.7"
emoji = "2.14.0"
faiss-cpu = "1.9.0"
filelock = "3.16.1"
flask = "3.0.3"
frozenlist = "1.4.1"
fsspec = "2024.9.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
huggingface-hub = "0.25.1"
idna = "3.10"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
joblib = "1.4.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
lxml = "5.3.0"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
mpmath = "1.3.0"
multidict = "6.1.0"
narwhals = "1.9.3"
networkx = "3.3"
numpy = "2.1.1"
packaging = "24.1"
pandas = "2.2.3"
pillow = "10.4.0"
propcache = "0.2.0"
protobuf = "5.28.2"
pyaes = "1.6.1"
pyarrow = "17.0.0"
pyasn1 = "0.6.1"
pydeck = "0.9.1"
pygments = "2.18.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
pyyaml = "6.0.2"
referencing = "0.35.1"
regex = "2024.9.11"
requests = "2.32.3"
rich = "13.9.2"
rpds-py = "0.20.0"
rsa = "4.9"
safetensors = "0.4.5"
scikit-learn = "1.5.2"
scipy = "1.14.1"
sentence-transformers = "3.1.1"
six = "1.16.0"
smmap = "5.0.1"
soupsieve = "2.6"
streamlit = "1.39.0"
sympy = "1.13.3"
telethon = "1.37.0"
tenacity = "9.0.0"
threadpoolctl = "3.5.0"
tokenizers = "0.20.0"
toml = "0.10.2"
torch = "2.4.1"
tornado = "6.4.1"
tqdm = "4.66.5"
transformers = "4.45.1"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
werkzeug = "3.0.4"
yarl = "1.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
